+++ CHANGES.rst
@@ -7,7 +7,9 @@ Bugs fixed
 * #12299: Defer loading themes defined via entry points until
   their explicit use by the user or a child theme.
   Patch by Adam Turner.
-
+* #12305: Return the default value for ``theme.get_config()`` with
+  an unsupported theme configuration section.
+  Patch by Adam Turner.
 
 Release 7.3.6 (released Apr 17, 2024)
 =====================================